Move everything inside "Transposition" class.
