+++ pwiz_tools/Shared/CommonUtil/SystemUtil/HashedObject.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
-namespace pwiz.Common.Storage
+namespace pwiz.Common.SystemUtil
 {
     public static class HashedObject
     {
